Moved components to MyMoney.Windows.Components namepace
