+++ MyMoney/MyMoney/MyMoney.csproj
@@ -100,15 +100,15 @@
     </Compile>
     <Compile Include="Controllers\DataController.cs" />
     <Compile Include="Windows\IView.cs" />
-    <Compile Include="Windows\User_Display\GraphHandler.cs" />
+    <Compile Include="Windows\Components\GraphHandler.cs" />
     <Compile Include="Windows\Main.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Windows\Main.Designer.cs">
       <DependentUpon>Main.cs</DependentUpon>
     </Compile>
-    <Compile Include="Windows\User_Display\ToolTipHandler.cs" />
-    <Compile Include="Windows\User_Display\TransactionView.cs" />
+    <Compile Include="Windows\Components\ToolTipHandler.cs" />
+    <Compile Include="Windows\Components\TransactionView.cs" />
     <Compile Include="Windows\MonthlyAllowanceChanger.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -119,7 +119,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Controllers\SQLController.cs" />
     <Compile Include="Model\Table\Table.cs" />
-    <Compile Include="Windows\User_Display\TransactionViewer.cs" />
+    <Compile Include="Windows\Components\TransactionViewer.cs" />
     <EmbeddedResource Include="Windows\AboutWindow.resx">
       <DependentUpon>AboutWindow.cs</DependentUpon>
     </EmbeddedResource>
