+++ MyMoney/MyMoney/Windows/Components/TransactionView.cs
@@ -2,60 +2,22 @@
 using MyMoney.Model.Table;
 using System.Windows.Forms;
 
-namespace MyMoney.Windows.User_Display
+namespace MyMoney.Windows.Components
 {
-    /// <summary>
-    /// An immutable object that encapsultes the components used to display a transaction on the main form.
-    /// </summary>
+
     public class TransactionView
     {
 
-        /// <summary>
-        /// The text box that will hold the date of the transaction.
-        /// </summary>
         public RichTextBox date;
 
-        /// <summary>
-        /// The text box that will hold the description of the transaction.
-        /// </summary>
         public RichTextBox description;
 
-        /// <summary>
-        /// The text box that will hold the amount of currency that 
-        /// was gained or lost through this transaction.
-        /// </summary>
         public RichTextBox amount;
 
-        /// <summary>
-        /// The button that will be used to delete this transaction.
-        /// </summary>
         public Button delete;
 
-        /// <summary>
-        /// Stores the data that will be displayed in this transacation view.
-        /// </summary>
         public Row row;
 
-        /// <summary>
-        /// Constructs a new transaction view.
-        /// </summary>
-        /// <param name="date">
-        /// The text box that will hold the date of the transaction.
-        /// </param>
-        /// <param name="description">
-        /// The text box that will hold the description of the transaction.
-        /// </param>
-        /// <param name="amount">
-        /// The text box that will hold the amount of currency that 
-        /// was gained or lost through this transaction.
-        /// </param>
-        /// <param name="delete">
-        /// The button that will be used to delete this transaction.
-        /// </param>
-        /// </param>
-        /// <param name="data">
-        /// The data that will be displayed in this transacation view.
-        /// </param>
         public TransactionView(RichTextBox date, RichTextBox description, RichTextBox amount, Button delete)
         {
             this.delete = delete;
@@ -66,10 +28,6 @@ namespace MyMoney.Windows.User_Display
 
         }
 
-        /// <summary>
-        /// Displays a specified <see cref="Row"/> in this <see cref="TransactionView"/>.
-        /// </summary>
-        /// <param name="row">The <see cref="Row"/> to be displayed.</param>
         public void setView(Row row)
         {
             // Set the row to be the same as the specified row.
@@ -108,12 +66,6 @@ namespace MyMoney.Windows.User_Display
             }
         }
 
-        /// <summary>
-        /// Updates the enabled state of a specified button to a specified bool state. 
-        /// This allows thread safe modification of the button.
-        /// </summary>
-        /// <param name="button">The button to be modified.</param>
-        /// <param name="enabled">The new enabled state of the button.</param>
         private void updateButton(Button button, bool enabled)
         {
             if (button.InvokeRequired)
@@ -129,13 +81,6 @@ namespace MyMoney.Windows.User_Display
             }
         }
 
-        /// <summary>
-        /// Updates the contents of a specified RichTextBox to a specified bool state. 
-        /// This allows thread safe modification of the RichTextBox. If the new contents 
-        /// are empty then the RichTextBox is disabled and vice versa if it is not.
-        /// </summary>
-        /// <param name="box">The RichTextBox to be modified.</param>
-        /// <param name="item">The new contents of the RichTextBox.</param>
         private void updateBox(RichTextBox box, string item)
         {
 
@@ -170,9 +115,6 @@ namespace MyMoney.Windows.User_Display
             }
         }
 
-        /// <summary>
-        /// Enables all the controls in the view.
-        /// </summary>
         public void enable()
         {
             date.Enabled = true;
@@ -181,9 +123,6 @@ namespace MyMoney.Windows.User_Display
             description.Enabled = true;
         }
 
-        /// <summary>
-        /// Disables all the contols in the view.
-        /// </summary>
         public void disable()
         {
             date.Enabled = false;
